requirements.txt
Flask[async]
requests
aiohttp
google-generativeai
beautifulsoup4
